Remove redundant selector and add profile form state
Removed the redundant selector in the `Profile` entity and integrated form state into the `Profile` state. This will allow easier updating of profile data through a form. Added a service for updating profile data, also added `readOnly` field to profile state to control the editability of data. Modified the `InfoItem` component to be able to handle changes of form fields. Also added handling for country and currency fields in `InfoItem`. Added useful translations and modified the `Input` ui component for handling read-only state.