Add Articles and ArticleDetails pages
The commit introduces two new pages: Articles and ArticleDetails. The sidebar link list is updated with a new item: articles that direct to the new Articles page. Lazy loading is added for the two new pages and their corresponding stylesheets and storybook files are created. The router configuration is updated to incorporate the new pages. Also, necessary import statements are adjusted and added. The locales files are updated to include '"page.articles": "Articles"' translations. The `publicPath: '/'` is added to `output` in webpack config to resolve the correct path in lazy loaded components.