Add fully functional 'comment form'
Added all necessary files for the correct work of new feature for users to comment on specific articles. This includes necessary slices, selectors, and async components for forms input. Action was taken to better the user's experience, round out the feature set with a Comment Form and increase user interaction. Now users can leave comments under an article, which also increases user engagement with the site.

Ensured test coverage and storybook integration for average and edge cases. Adjusted fetch function in Profile's 'fetchProfileData' to take the user id.

Styling for form was defined in separate SCSS file for maintainability. The form includes an input field and button.