Correct styling, update story decorators, remove redundant code
Adjusted the style syntax in dark.scss to fix the incorrect multi-line declaration for the card-border-hover-color. Updated StoreDecorator usage in the story definition for AddCommentForm, adding initial state for the addCommentForm and updating the pathing due to a file rename. Removed unnecessary eslint disable line in StoreDecorator.tsx file. These changes are necessary for improved code styling, better support for storybook stories, and to remove superfluous code.