Refactor Article components and implemented infinite scrolling
In this commit, a significant refactor was made to the components related to rendering articles. One of the changes was to omit the unused `index` in the `map` function in the `ArticleTextBlockComponent`, ensuring cleaner, more efficient code.

Infinite scrolling was also implemented in the `ArticleViewSelector` for improved user experience, especially when dealing with long lists of articles. Each `viewType` is now assigned a unique key, which enhances React's handling and performance of these elements.

Other changes related to 'initArticlesPage.ts', login components and 'useDynamicModuleLoad.tsx' were made as a part of the code refactoring and optimization process for the article-related functionalities.