Set up request interceptor to add authorization header
Refactored the way the authorization header is added to the API calls. Instead of setting it directly in the API instance, a new request interceptor is introduced. The interceptor is designed to check if headers are present in the configuration, then add the 'authorization' field picked from local storage. This ensures that the latest token is always used for authenticated requests.