Remove unused class styles in SideBarItem.module.scss
The unused styles in SideBarItem.module.scss were removed in order to optimize the SideBarItem and to make the widget layout cleaner. This cleanup was done as part of a larger refactor where Flex component is being used everywhere for layout, making several .scss files unnecessary. By doing this, it also helped in reducing duplicate code and increasing code maintainability.