Fixing formatting in Dropdown module SCSS file
In this commit, we are improving the code readability by fixing the formatting in the Dropdown.module.scss file. The color attribute had unwanted space before the ending semicolon, which is now removed to ensure consistent code formatting across inline styling attributes. While this doesn't add to functionality changes, it helps in maintaining a neater and consistent codebase.